chore: small changes from tests

+++ src/handlers/update-base-rate.ts
@@ -41,7 +41,6 @@ export async function updateBaseRate(context: Context) {
       continue;
     }
     await updateLabelsFromBaseRate(context, repo.owner.login, repo.name, labels, rates);
-    logger.info(`Updated labels for ${repo.name}`);
   }
 }
 
@@ -62,7 +61,6 @@ async function updateLabelsFromBaseRate(context: Context, owner: string, repo: s
     },
   } = context;
   const { previousBaseRate, newBaseRate } = rates;
-
   const timeLabels = labels.filter((label) => label.name.includes("Time:"));
   const priorityLabels = labels.filter((label) => label.name.includes("Priority:"));
   let priceLabels: Label[] | null = labels.filter((label) => label.name.includes("Price:"));
@@ -72,12 +70,16 @@ async function updateLabelsFromBaseRate(context: Context, owner: string, repo: s
   }
 
   if (!newBaseRate) {
-    throw logger.error("No new base rate found");
+    return;
   }
 
   const uniqueTimeLabels = [...new Set(timeLabels.map((label) => label.name).concat(time))];
   const uniquePriorityLabels = [...new Set(priorityLabels.map((label) => label.name).concat(priority))];
   const repoIssues: Awaited<ReturnType<typeof fetchIssuesWithTandP>> | null = await fetchIssuesWithTandP(context, owner, repo);
+
+  if (!repoIssues.length) {
+    return;
+  }
 
   if (hasAssistivePricing) {
     await deleteAllPriceLabels(context, owner, repo, priceLabels);
